backend/requirements.txt
# Python dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
Pillow==10.2.0
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
numpy==1.26.3
